(PC-14127) fix(Firebase8): refacto to universal interface
