+++ src/features/maintenance/MaintenanceErrorPage.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 
 import { Maintenance } from 'features/maintenance/Maintenance'
 import { useMaintenance } from 'features/maintenance/useMaintenance'
-import { MAINTENANCE } from 'libs/firestore/maintenance'
+import { MAINTENANCE } from 'libs/firebase/firestore/maintenance'
 
 type Props = {
   resetErrorBoundary: () => void
