--- src/libs/firebase/index.ts
@@ -1,1 +1,0 @@
-export { getFirebaseApp } from './firebase'
